+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - `Entities` support `__getitem__` with uuds [#159](https://github.com/ie3-institute/pypsdm/pull/159)
 - Support for lines and transformer admittance matrix calculation [#161](https://github.com/ie3-institute/pypsdm/pull/161)
 - `ExtendedNode(s)Result` are now calculated using the grids admittance matrix 
+- Specific `Line(s)Result` and line utilisation calculation [[#176](https://github.com/ie3-institute/pypsdm/issues/176)]
 
 ### Changed
 - Renamed `EnhancedNodeResult` and `EnhancedNodesResult` to `ExtendedNoderesult` and `ExtendedNodesresult` [#155](https://github.com/ie3-institute/pypsdm/pull/155)
